requirements.txt
category_encoders==2.5.1.post0
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.3.0
plotly==5.9.0
scikit_learn==1.1.3
seaborn==0.12.1
streamlit==1.14.0
xgboost==1.6.2
